requirements.txt
Flask==3.0.0
APScheduler==3.10.4
requests==2.31.0
peewee==3.17.0
python-dotenv==1.0.0
